+++ app/src/main/java/com/wallet/crypto/alphawallet/viewmodel/SellDetailModel.java
@@ -55,7 +55,7 @@ public class SellDetailModel extends BaseViewModel {
         defaultWallet.setValue(wallet);
     }
 
-    public void generateSalesOrders(String contractAddr, BigInteger price, short[] ticketIndicies, int firstTicketId)
+    public void generateSalesOrders(String contractAddr, BigInteger price, int[] ticketIndicies, int firstTicketId)
     {
         marketQueueService.createSalesOrders(defaultWallet.getValue(), price, ticketIndicies, contractAddr, firstTicketId);
     }
